change btc_balance integer to float

+++ db/migrate/20180818001942_create_profits.rb
@@ -3,7 +3,7 @@ class CreateProfits < ActiveRecord::Migration[5.2]
     create_table :profits do |t|
       t.text :date
       t.integer :total_jpy_balance
-      t.integer :total_btc_balance
+      t.float :total_btc_balance
       t.integer :average_btc_price
       t.integer :total_balance
       t.integer :profit
